requirements.txt
streamlit
tensorflow-cpu
scikit-learn
pandas
numpy
joblib
keras-tuner